[FIND-269] changed numeric restriction type from long to double and added migration scripts. [rev:alex.scown, matthew.gordon2]

+++ core/src/main/java/com/hp/autonomy/frontend/find/core/savedsearches/ParametricRange.java
@@ -20,8 +20,8 @@ import javax.persistence.Enumerated;
 @NoArgsConstructor
 public class ParametricRange {
     private String field;
-    private long min;
-    private long max;
+    private double min;
+    private double max;
     @Enumerated(EnumType.ORDINAL)
     private Type type;
 
@@ -34,8 +34,8 @@ public class ParametricRange {
     @JsonCreator
     public ParametricRange(
             @JsonProperty("field") final String field,
-            @JsonProperty("min") final long min,
-            @JsonProperty("max") final long max,
+            @JsonProperty("min") final double min,
+            @JsonProperty("max") final double max,
             @JsonProperty("type") final Type type
     ) {
         this.field = field;
